@@ -7,14 +7,14 @@
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
     "clarifai": "^2.9.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-particles-js": "^3.4.1",
+    "particles-bg": "^2.5.5",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-parallax-tilt": "^1.7.77",
     "react-scripts": "^5.0.1",
-    "react-tilt": "^0.1.4",
     "serve": "^14.1.2",
     "tachyons": "^4.12.0",
-    "web-vitals": "^1.1.1"
+    "web-vitals": "^3.1.0"
   },
   "scripts": {
     "start": "serve -s build",
